requirements.txt
pandas
reportlab
pillow
pymupdf